feat: relocated the result functions to the proper package
