+++ clients/camunda-spring-boot-starter/src/main/java/io/camunda/client/jobhandling/JobHandlerInvokingBeans.java
@@ -23,8 +23,10 @@ import io.camunda.client.api.worker.JobClient;
 import io.camunda.client.api.worker.JobHandler;
 import io.camunda.client.impl.Loggers;
 import io.camunda.client.jobhandling.parameter.ParameterResolver;
+import io.camunda.client.jobhandling.result.AdHocSubProcessResultFunction;
 import io.camunda.client.jobhandling.result.ResultProcessor;
 import io.camunda.client.jobhandling.result.ResultProcessorContext;
+import io.camunda.client.jobhandling.result.UserTaskResultFunction;
 import io.camunda.client.metrics.JobHandlerMetrics;
 import io.camunda.client.metrics.MetricsRecorder;
 import io.camunda.client.metrics.MetricsRecorder.CounterMetricsContext;
