+++ clients/camunda-spring-boot-starter/src/main/java/io/camunda/client/jobhandling/result/AdHocSubProcessResultFunction.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.client.jobhandling;
+package io.camunda.client.jobhandling.result;
 
 import io.camunda.client.api.command.CompleteAdHocSubProcessResultStep1;
 import io.camunda.client.api.command.CompleteJobResult;
