+++ clients/camunda-spring-boot-starter/src/main/java/io/camunda/client/jobhandling/result/UserTaskResultFunction.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.client.jobhandling;
+package io.camunda.client.jobhandling.result;
 
 import io.camunda.client.api.command.CompleteJobResult;
 import io.camunda.client.api.command.CompleteUserTaskJobResultStep1;
